wat the description says...
Had real progress. Player was 'collecting' (but the items stay on the map) and updating the counters. Sound effects were present in collecting coins and cans, dying and restarting the game (still is as for the re/start of game). Life was evidently getting subtracted when falling from a great height. Timer was working (still is). But now, the player is just stuck on a landing animation loop and is locking inputs and unabling the 'collection' of items. Should've saved it when i was ok :'>